# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Rare District — Nigerian fashion marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegistration"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /auth/login:
    post:
      operationId: loginUser
      tags: [Auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLogin"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [Auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /auth/me/profile:
    patch:
      operationId: updateMyProfile
      tags: [Auth]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /auth/google/callback:
    get:
      operationId: googleAuthCallback
      tags: [Auth]
      summary: Google OAuth callback
      parameters:
        - name: code
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
      responses:
        "302":
          description: Redirect after auth
        "200":
          description: Auth success with token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"

  # ── VENDORS ───────────────────────────────────────────────────────────────
...
  /coupons/validate:
    post:
      operationId: validateCoupon
      tags: [Coupons]
      summary: Validate a coupon code against an order total
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponValidation"
      responses:
        "200":
          description: Coupon validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CouponResult"
        "400":
          description: Invalid or expired coupon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /coupons:
    get:
      operationId: listCoupons
      tags: [Coupons]
      summary: List all coupons (admin only)
      responses:
        "200":
          description: Coupons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Coupon"

    post:
      operationId: createCoupon
      tags: [Coupons]
      summary: Create a coupon (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponInput"
      responses:
        "201":
          description: Created coupon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
  /coupons/{id}:
    patch:
      operationId: updateCoupon
      tags: [Coupons]
      summary: Update a coupon (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponUpdate"
      responses:
        "200":
          description: Updated coupon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"

    delete:
      operationId: deleteCoupon
      tags: [Coupons]
      summary: Delete a coupon (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── REFERRALS ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    User:
      type: object
      required: [id, email, role, referralCode, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [shopper, vendor, admin]
        referralCode:
          type: string
        referredBy:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    UserRegistration:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 6
        name:
          type: string
          minLength: 1
        role:
          type: string
          enum: [shopper, vendor]
          default: shopper
        referralCode:
          type: string
    UserLogin:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    UserProfileUpdate:
      type: object
      properties:
        name:
          type: string
        avatarUrl:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # ── Vendor ──────────────────────────────────────────────────────────────
...
    Coupon:
      type: object
      required: [id, code, type, value, isActive, createdAt]
      properties:
        id:
          type: integer
        code:
          type: string
        type:
          type: string
          enum: [percentage, fixed]
        value:
          type: number
        minOrderAmount:
          type: ["number", "null"]
        maxUses:
          type: ["integer", "null"]
        usedCount:
          type: integer
        expiresAt:
          type: ["string", "null"]
          format: date-time
        isActive:
          type: boolean
        isReferral:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CouponInput:
      type: object
      required: [code, type, value]
      properties:
        code:
          type: string
        type:
          type: string
          enum: [percentage, fixed]
        value:
          type: number
        minOrderAmount:
          type: number
        maxUses:
          type: integer
        expiresAt:
          type: string
          format: date-time
    CouponUpdate:
      type: object
      properties:
        value:
          type: number
        minOrderAmount:
          type: number
        maxUses:
          type: integer
        expiresAt:
          type: string
          format: date-time
        isActive:
          type: boolean
    CouponValidation:
      type: object
      required: [code, orderTotal]
      properties:
        code:
          type: string
        orderTotal:
          type: number
    CouponResult:
      type: object
      required: [valid, discountAmount, finalTotal]
      properties:
        valid:
          type: boolean
        discountAmount:
          type: number
        finalTotal:
          type: number
        coupon:
          $ref: "#/components/schemas/Coupon"

    # ── Referrals ─────────────────────────────────────────────────────────────
...